Create a custom hook for add items to shoppingList
